requirements.txt
pymysql>=1.1.0
requests>=2.32.0
beautifulsoup4>=4.12.0
python-dateutil>=2.9.0.post0
playwright>=1.52.0
